+++ SexyAppFramework/SysFont.cpp
@@ -172,7 +172,7 @@ int	SysFont::StringWidth(const SexyString& theString)
 #endif
 	{
 		RECT aRect = {0, 0, 0, 0};	
-		DrawTextEx(aDC, (SexyChar*)theString.c_str(), theString.length(), &aRect, DT_CALCRECT | DT_NOPREFIX, NULL);
+		DrawTextEx(aDC, (SexyChar*)theString.c_str(), (int)theString.length(), &aRect, DT_CALCRECT | DT_NOPREFIX, NULL);
 		aWidth = aRect.right;
 	}
 
@@ -208,14 +208,14 @@ void SysFont::DrawString(Graphics* g, int theX, int theY, const SexyString& theS
 				if (mDrawShadow)
 				{
 					SetTextColor(aDC, RGB(0,0,0));
-					TextOut(aDC, theX + g->mTransX+1, theY - mAscent + 1 + g->mTransY+1, theString.c_str(), theString.length());
+					TextOut(aDC, theX + g->mTransX+1, theY - mAscent + 1 + g->mTransY+1, theString.c_str(), (int)theString.length());
 					if (mSimulateBold)
-						TextOut(aDC, theX + g->mTransX+2, theY - mAscent + 1 + g->mTransY+1, theString.c_str(), theString.length());
+						TextOut(aDC, theX + g->mTransX+2, theY - mAscent + 1 + g->mTransY+1, theString.c_str(), (int)theString.length());
 				}
 				SetTextColor(aDC, RGB(theColor.GetRed(), theColor.GetGreen(), theColor.GetBlue()));
-				TextOut(aDC, theX + g->mTransX, theY - mAscent + 1 + g->mTransY, theString.c_str(), theString.length());
+				TextOut(aDC, theX + g->mTransX, theY - mAscent + 1 + g->mTransY, theString.c_str(), (int)theString.length());
 				if (mSimulateBold)
-					TextOut(aDC, theX + g->mTransX + 1, theY - mAscent + 1 + g->mTransY, theString.c_str(), theString.length());
+					TextOut(aDC, theX + g->mTransX + 1, theY - mAscent + 1 + g->mTransY, theString.c_str(), (int)theString.length());
 
 				::SelectObject(aDC, anOldFont);
 				aSurface->ReleaseDC(aDC);
